change(driver): rename to observability
